DBScale/requirement.txt
asgiref==3.5.0
Django==3.2.9
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
